sdks/python/requirements-dev.txt
# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
requests-mock>=1.11.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
